requirements.txt
openai
openai-whisper
SpeechRecognition
PyAudio
streamlit 
streamlit-chat
pyyaml